When an event occurs, sometimes you need to detect it
Things that occur are events.

Gestures are things that occur.

Tapping something is a type of gesture.

Tapping something is an event.

You will not be satisfied if you look at the screen as an indicator that your event was detected.

No. This time, you will need to look in the console for your proof.