Frequently in life, you may need to count some things
With this hot new app, you can press things to make the count go up or
down depending on how you feel.

As far as computer programming goes, the sophistication here is that you
can write some functions that will increase or decrease a number, and
you can trigger the updated number to be displayed to the user.